BUG FIX : in the login page (IAE-33) The login page is not filtered anymore by the IntactSessionRequestFilter.
The LoginAction call to the authenticate method of the AuthenticateUser class. This method load the driver class and with the database connection Url, the login name and password it try to get a connection to the database and to get the metaData.
If this fails it send an AuthenticateException. The web.xml is configured so that those Exception are handled and nicely displayed ('Wrong login or password' type of page).
[ Show » ]
Catherine Leroy [03-Apr-07 14:45] The login page is not filtered anymore by the IntactSessionRequestFilter. The LoginAction call to the authenticate method of the AuthenticateUser class. This method load the driver class and with the database connection Url, the login name and password it try to get a connection to the database and to get the metaData. If this fails it send an AuthenticateException. The web.xml is configured so that those Exception are handled and nicely displayed ('Wrong login or password' type of page).