BadParticipantFoundEvent and BadParticipantFixer : clone a protein and attach bad ranges to it. We had 'no-uniprot-update' and a caution
Duplicate : before merging two proteins with different sequences, shift the ranges. If not possible, don't merge the duplicate and add 'no-uniprot-update' and a caution.

Added the new event and listener in ProteinUpdateProcessor and in the writer listener.

Changed the range checker to separate the invalid range process from the range shifting. It will not consider as a bad range if after shifting the range positions, the feature is not at the end/beginning of the sequence but will add a caution.

Deleted the methods processDuplicate and processDuplicateTranscript of the ProteinServiceImpl because it is a job only done by the DuplicateFixer and DuplicateFinder.

Added several tests but need to add a final test for the newest listener.