BadParticipantFixer is now called OutOfDateParticipantFixer and BadParticipantFoundEvent is now called OutOfDateParticipantFoundEvent
Fixed a general bug : when a protein contains several uniprot identities, merge the identical ones and add an error if two different uniprot identities are found.

In ProteinServiceImpl, if there are several uniprot entries matching the uniprot ac, we check first if it is not a protein transcript. If yes, it means that the protein transcript has several parents and can be updated