ProteinEvent : now contains a variable uniprotPrimaryId which is the uniprot identity of the protein. Contains also a variable uniprotProtein which is the uniprot entry matching the protein in intact.
UpdateError : an enum to list all possible type of error during the update

UpdateErrorEvent : when an erro occur, we fire this event to log in a separate file the detail of the errors

ProteinProcessor : commented the line to not add the DeadUniprotListener which is leading to a lot of broken tests for the moment.

ReportWriterListener : will write in a separate file the errors and the secondary proteins updated

DuplicateFinder and DuplicateFixer : fixed a bug : the ranges were never shifted before the merge

UniprotPrimaryAcUpdater : this listener is in charge to update all secondary proteins (update the uniprot identity)

UniprotProteinRetrieverListener : this listener will now be in charge to retrieve the uniprot entry matching the protein. 

UniprotUpdateFilterListener : this listener is charged to filter during the pre-process event the no-uniprot update and the proteins without oir with several uniprot identity(ies)

UniprotProteinUpdater : this listener is not extending ProteinServiceImpl anymore and is only updating the proteins using the uniprot entry. It is only working during the process phase of the uniprot update.

Started to add several tests but it is not finished