UpdaeCaseEvent : Collection<? extends Protein> becomes Collection<Protein>
UpdateError : added a few new errors

DuplicateFinder and DuplicateFixer : is not a listener anymore. fixed a few bugs and split the methods into three methods : proteins, isoforms and feature chains

ProteinDeleter : is not a listener anymore. now fireOnDelete before deleting the proteins. So the proteins deleted are logged only if they are deleted

OutOfDateParticipantFoundEvent : extends UpdateCaseEvent because we need to have the list of isoforms and feature chains to be able to remap to an existing splice variant if necessary.

UpdateCaseEvent : instead of a list of Proteins for the splice variants and feature chains, it is a list of ProteinTranscript which contains both the protein in IntAct and the protein transcript in uniprot

UpdateError : Changed one error

AbstractProteinUpdateProcessorListener : does not implement ProcessorListener because we don't use pre-process and process events

LoggingProcessorListener : does not implement ProcessorListener because we don't use pre-process and process events

ReportWriterListener : does not implement ProcessorListener because we don't use pre-process and process events

actions package : contains all the previous listeners which are not listeners anymore but just actions which need to be done in a specific order (DuplicateFixer, DuplicateFinder, OutOfDateParticipantFixer, ProteinDeleter, ProteinUpdateFilter, ProtWithoutInteractionDeleter, RangeFixer, UniprotIdentityUpdater, UniprotProteinRetriever and UniprotProteinUpdater)

DuplicateReport : contains the results of the merge (original protein, map containing proteins with range conflicts and the affected components)

ProteinProcessor : run action by action, doesn't have a list of ProteinprocessorListener but ProteinUpdateProcessorListener

ProteinUpdateProcessir : updated the list of listeners

Fixed several failing tests