Use isFeatureChain method of ProteinUtils to check if a protein is a feature chain.
UniprotProteinretrieverImpl : return the remapped uniprot protein when the remapping is successful.

ProteinUpdateFilter : return the remapped uniprot identity when remapping successful.

If the remapped protein is not attached to the same uniprot entry currently updated, it nwill be removed from the list of protein processed and updated for the current uniprot entry and will be updated later