Bug fixed : the method setAnnotatedObject of the different controllers returned NullPointerException when deleting proteins for instance (interactorController). Added some checking before setting the ac in each controller.
Refactoring : the protein transcripts created are now marked as CREATED_TRANSCRIPTS and the master protein will be created when the persisterController saveAll or in the doSave method of AnnotatedObjectController so when saving an interaction, we don't need to go through the all list of participants. Heavy when having more than 20 components

Bug Fixed : in the saveAll method of persisterController, removed the @Transaction annotation because the corePersister already opens a transaction and can causes TransactionExceptions : a transaction is already opened. This was a problem when creating the master proteins where some object were then detached after creating the master protein. Removing this annotation does not stop deleting, creating or updating intactObjects.