Added a property isCheckInitializedCollections for the converters to give the option to check if collections are initialized in hibernate or not.
Doing this check everytime is considerably slowing the process. If we are in a transaction and hibernate session is open (and object is in the hibernate session), we should never check if collections are initialized because they will not be. In this case, we should trust Hibernate to fetch the objects and not by-pass it using the ensureInitialized methods of IntactCore. For retrocompatibility, the isCheckInitializedCollections is true by default but it can be changed.

Did some optimization